+++ tutorials.rst
@@ -1,0 +1,4 @@
+Tutorials
+*********
+
+The following pages contain the TARDIS tutorials:
